fix quotes input
