+++ src/pages/MyProfile/CharCard.jsx
@@ -1,55 +1,25 @@
-import React, { useState } from "react";
-import { addQuoteToProfile } from "../../services/characterService";
+import React from "react";
 import QuoteCard from "./QuoteCard";
+import AddQuote from "./AddQuote";
 
-const CharCard = ({character, updateCharacter, handleDeleteCharacter}) => {
-  const [content, setContent] = useState("")
-
-  const formData = {
-    content: content,
-  };
-
-  const addQuote = async (e, id) => {
-    e.preventDefault();
-    try {
-      const updatedChar = await addQuoteToProfile(id, formData);
-      updateCharacter(updatedChar, id);
-      setContent("");
-    } catch (error) {
-      throw error;
-    }
-  };
-
+const CharCard = ({ character, updateCharacter, handleDeleteCharacter }) => {
   const randomQuote =
     character.quotes[Math.floor(Math.random() * character.quotes.length)];
 
   return (
     <div>
-      <form onSubmit={(e) => addQuote(e, character._id)}>
-        <p>{character.name}</p>
-        <img
-          src={`${character.thumbnail.path}.${character.thumbnail.extension}`}
-          alt=""
-          height="300px"
-          width="300px"
-        ></img>
-        <input
-          type="text"
-          required
-          name="content"
-          autoComplete="off"
-          placeholder="Say something!"
-          value={content}
-          onChange={(e) => setContent(e.target.value)}
-        />
-        <button type="submit">Add Quote</button>
-      </form>
-      <button onClick={()=> handleDeleteCharacter(character._id)}>Delete</button>
-        <QuoteCard
-        characterId={character._id}
-        randomQuote={randomQuote}
-        />
-        
+      <p>{character.name}</p>
+      <img
+        src={`${character.thumbnail.path}.${character.thumbnail.extension}`}
+        alt=""
+        height="300px"
+        width="300px"
+      ></img>
+      <AddQuote character={character} updateCharacter={updateCharacter} />
+      <button onClick={() => handleDeleteCharacter(character._id)}>
+        Delete Character
+      </button>
+      <QuoteCard characterId={character._id} randomQuote={randomQuote} />
     </div>
   );
 };
