+++ src/pages/Signup/Signup.module.css
@@ -1,19 +1,25 @@
+.leftcontainer {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
 
 
-.leftcontainer{
+.container{
+  color: rgb(211, 0, 0);
+  background-color: rgba(20, 2, 2, 0);
+}
+
+/* .leftcontainer{
 position:relative;
 bottom:22rem;
 margin-left: 90px;
 z-index: -1;
-}
+} */
 
-.container{
-  color:rgb(211, 0, 0);
-  background-color: rgba(20, 2, 2, 0);
-}
-.signup{
+/* .signup{
   margin-left: 70rem;
-}
+} */
 
 /* .bubble {
   color: red;
